fixed bags

+++ src/4_hw_goverments/queue.c
@@ -32,13 +32,12 @@ void heapPush(Heap* heap, int dist, int city, int state)
 {
     if (heap->size == heap->capacity) {
         size_t newCapacity = heap->capacity * 2;
-        // Используем временную переменную, чтобы избежать утечки при ошибке realloc
         QueueNode* newData = (QueueNode*)realloc(heap->data, newCapacity * sizeof(QueueNode));
         if (newData == NULL) {
             return;
         }
         heap->data = newData;
-        heap->capacity = newCapacity;
+        heap->capacity = (int)newCapacity;
     }
     int idx = heap->size++;
     heap->data[idx].distForCapital = dist;
@@ -46,7 +45,6 @@ void heapPush(Heap* heap, int dist, int city, int state)
     heap->data[idx].numGoverment = state;
     heapUp(heap, idx);
 }
-
 void heapDown(Heap* heap, int idx)
 {
     int size = heap->size;
